Replace cdp{cmd,event,type} with separated domains
NOTE: Breaking change. Instead of domain types belonging to one of
cdpcmd, cdpevent or cdptype, they all now reside in a single package,
e.g. `network`.

This is a big refactor of the generated code that represents all the
different CDP domains.

For the common use-case, only the args need to be rewritten:

1. `cdpcmd.NewNetworkEnableArgs()` -> `network.NewEnableArgs()`
2. `&cdpcmd.NetworkEnableArgs{}` -> `&network.EnableArgs{}`